use the openinference core typeguards in a couple more places, extract some shared helper methods from converse and invoke request extraction, update the semantic convention LLM System enum and use it in the helper
