+++ js/packages/openinference-instrumentation-bedrock/test/instrumentation.test.ts
@@ -1809,7 +1809,7 @@ At the top of the",
   "llm.output_messages.0.message.role": "assistant",
   "llm.provider": "aws",
   "llm.stop_reason": "end_turn",
-  "llm.system": "mistral",
+  "llm.system": "mistralai",
   "llm.token_count.completion": 94,
   "llm.token_count.prompt": 18,
   "llm.token_count.total": 112,
